feat: fetch comments using graphql
